feat(editor): Make the left sidebar in Expressions editor draggable (#11838)
