+++ packages/editor-ui/src/components/ExpressionEditModal.vue
@@ -23,6 +23,10 @@ import { dropInExpressionEditor } from '@/plugins/codemirror/dragAndDrop';
 
 import { APP_MODALS_ELEMENT_ID } from '@/constants';
 import { N8nInput, N8nText } from 'n8n-design-system';
+import { N8nResizeWrapper, type ResizeData } from 'n8n-design-system';
+import { useThrottleFn } from '@vueuse/core';
+
+const DEFAULT_LEFT_SIDEBAR_WIDTH = 360;
 
 type Props = {
 	parameter: INodeProperties;
@@ -56,6 +60,7 @@ const { debounce } = useDebounce();
 const segments = ref<Segment[]>([]);
 const search = ref('');
 const appliedSearch = ref('');
+const sidebarWidth = ref(DEFAULT_LEFT_SIDEBAR_WIDTH);
 const expressionInputRef = ref<InstanceType<typeof ExpressionEditorModalInput>>();
 const expressionResultRef = ref<InstanceType<typeof ExpressionOutput>>();
 const theme = outputTheme();
@@ -122,6 +127,12 @@ async function onDrop(expression: string, event: MouseEvent) {
 
 	await dropInExpressionEditor(toRaw(inputEditor.value), event, expression);
 }
+
+function onResize(event: ResizeData) {
+	sidebarWidth.value = event.width;
+}
+
+const onResizeThrottle = useThrottleFn(onResize, 10);
 </script>
 
 <template>
@@ -136,28 +147,37 @@ async function onDrop(expression: string, event: MouseEvent) {
 			<Close height="18" width="18" />
 		</button>
 		<div :class="$style.container">
-			<div :class="$style.sidebar">
-				<N8nInput
-					v-model="search"
-					size="small"
-					:class="$style.search"
-					:placeholder="i18n.baseText('ndv.search.placeholder.input.schema')"
-				>
-					<template #prefix>
-						<N8nIcon :class="$style.ioSearchIcon" icon="search" />
-					</template>
-				</N8nInput>
+			<N8nResizeWrapper
+				:width="sidebarWidth"
+				:min-width="200"
+				:style="{ width: `${sidebarWidth}px` }"
+				:grid-size="8"
+				:supported-directions="['left', 'right']"
+				@resize="onResizeThrottle"
+			>
+				<div :class="$style.sidebar">
+					<N8nInput
+						v-model="search"
+						size="small"
+						:class="$style.search"
+						:placeholder="i18n.baseText('ndv.search.placeholder.input.schema')"
+					>
+						<template #prefix>
+							<N8nIcon :class="$style.ioSearchIcon" icon="search" />
+						</template>
+					</N8nInput>
 
-				<RunDataSchema
-					:class="$style.schema"
-					:search="appliedSearch"
-					:nodes="parentNodes"
-					:mapping-enabled="!isReadOnly"
-					:connection-type="NodeConnectionType.Main"
-					pane-type="input"
-					context="modal"
-				/>
-			</div>
+					<RunDataSchema
+						:class="$style.schema"
+						:search="appliedSearch"
+						:nodes="parentNodes"
+						:mapping-enabled="!isReadOnly"
+						:connection-type="NodeConnectionType.Main"
+						pane-type="input"
+						context="modal"
+					/>
+				</div>
+			</N8nResizeWrapper>
 
 			<div :class="$style.io">
 				<div :class="$style.input">
@@ -239,7 +259,7 @@ async function onDrop(expression: string, event: MouseEvent) {
 .container {
 	display: flex;
 	flex-flow: row nowrap;
-	gap: var(--spacing-s);
+	gap: var(--spacing-2xs);
 	height: 100%;
 }
